pyproject.toml
[project]
name = "jxr-box"
version = "0.1.0"
description = ""
authors = [
    {name = "Finimen",email = "finimensniper@gmail.com"}
]
readme = "README.md"
packages = [{include = "scr"}]
requires-python = ">=3.14"
dependencies = [
    "redis (>=7.4.0,<8.0.0)",
    "celery (>=5.6.3,<6.0.0)",
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.43.0,<0.44.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "apscheduler (>=3.11.2,<4.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "hiredis (>=3.3.1,<4.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "ruff (>=0.15.9,<0.16.0)",
    "mypy (>=1.20.0,<2.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--strict-markers", "--tb=short", "--disable-warnings", "--asyncio-mode=auto"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "asyncio: Async tests",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
